zsh completions: do not group options if zsh < 5.4
The `+ group` syntax to group options passed on to _arguments was
only added in zsh 5.4. This was causing an error when using the
completion script on zsh < 5.4. This commit fixes this by attempting
to obtain the zsh version by running `zsh --version`. zsh < 5.4 is
assumed upon failure to be on the safe side.