No longer call asdict in ConfigBase.to_dict
If an entry was itself a dataclass and specified a custom to_toml,
the call to asdict would recursively cast the dataclass to a dict
and call the custom to_toml on that dict instead of the dataclass
instance.  This commit fixes this by avoiding the call to asdict
altogether.